extract logic into custom hook
