+++ src/components/Slider.tsx
@@ -1,15 +1,9 @@
-import {
-  useRef,
-  useState,
-  useLayoutEffect,
-  MouseEvent,
-  TouchEvent,
-} from "react";
 import { Link } from "react-router-dom";
 
 import smallArrowRight from "../assets/arrow-right.svg";
 import arrowRight from "../assets/image-arrow-right.svg";
 import arrowLeft from "../assets/image-arrow-left.svg";
+import { useSlider } from "../hooks/useSlider";
 
 export type SliderElement = {
   title: string;
@@ -23,100 +17,25 @@ type SliderProps = {
 };
 
 export function Slider({ value }: SliderProps) {
-  const carouselRef = useRef<null | HTMLDivElement>(null);
-  const arrowRightRef = useRef<null | HTMLDivElement>(null);
-  const arrowLeftRef = useRef<null | HTMLDivElement>(null);
-
-  const [isDragStart, setIsDragStart] = useState(false);
-  const [isDragging, setIsDragging] = useState(false);
-
-  const [prevPageX, setPrevPageX] = useState(0);
-  const [prevScrollLeft, setPrevScrollLeft] = useState(0);
-  const [positionDiff, setPositionDiff] = useState(0);
-
-  const showHideIcons = () => {
-    // showing and hiding prev/next icon according to carousel scroll left value
-    if (carouselRef.current && arrowRightRef.current && arrowLeftRef.current) {
-      let scrollWidth: number =
-        carouselRef.current.scrollWidth - carouselRef.current.clientWidth;
-
-      arrowRightRef.current.style.display =
-        carouselRef.current.scrollLeft == scrollWidth ? "none" : "block";
-      arrowLeftRef.current.style.display =
-        carouselRef.current.scrollLeft == 0 ? "none" : "block";
-    }
-  };
-
-  const handleMouseDown = (event: MouseEvent<HTMLDivElement>) => {
-    setIsDragStart(true);
-    setPrevPageX(event.pageX); // horizontal coordinate of mouse click relative to the left edge of parent element
-    if (carouselRef.current) {
-      setPrevScrollLeft(carouselRef.current.scrollLeft); // current scroll starting from left of selected element
-    }
-  };
-
-  const handleMouseMove = (event: MouseEvent<HTMLDivElement>) => {
-    if (!carouselRef.current) return;
-    event.preventDefault();
-
-    // when we stop holding mouse button stop dragging
-    if (!isDragStart) {
-      return;
-    }
-    setIsDragging(true);
-
-    // scroll when moving mouse
-    setPositionDiff(event.pageX - prevPageX);
-    carouselRef.current.scrollLeft = prevScrollLeft - positionDiff;
-  };
-
-  const handleMouseUp = () => {
-    setIsDragStart(false);
-    if (!isDragging) return;
-    setIsDragging(false);
-
-    // cleanup
-    setPositionDiff(0);
-  };
-
-  function handleTouchDown(event: TouchEvent<HTMLDivElement>) {
-    setIsDragStart(true);
-    setPrevPageX(event.touches[0].pageX);
-    if (carouselRef.current) {
-      setPrevScrollLeft(carouselRef.current.scrollLeft);
-    }
-  }
-
-  const handleTouchMove = (event: TouchEvent<HTMLDivElement>) => {
-    if (!carouselRef.current) return;
-    event.preventDefault();
-    if (!isDragStart) {
-      return;
-    }
-    setIsDragging(true);
-    setPositionDiff(event.touches[0].pageX - prevPageX);
-    carouselRef.current.scrollLeft = prevScrollLeft - positionDiff;
-  };
-
-  function scrollBy(value: number, direction: string): void {
-    if (!carouselRef.current) return;
-
-    if (direction === "right") {
-      carouselRef.current.scrollLeft += value;
-    } else if (direction === "left") {
-      carouselRef.current.scrollLeft -= value;
-    }
-  }
-
-  useLayoutEffect(() => {
-    showHideIcons();
-  });
+  const {
+    isDragStart,
+    carouselRef,
+    arrowRightRef,
+    arrowLeftRef,
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleTouchDown,
+    handleTouchMove,
+    scrollLeftBy,
+    scrollRightBy,
+  } = useSlider();
 
   return (
     <div className="featured__right">
       <div
         ref={carouselRef}
-        className={`featured__carousel ${isDragStart ? "dragging" : ""}`}
+        className={`featured__carousel ${isDragStart && "dragging"}`}
         onMouseDown={(e) => handleMouseDown(e)}
         onMouseMove={(e) => handleMouseMove(e)}
         onMouseUp={() => handleMouseUp()}
@@ -129,7 +48,7 @@ export function Slider({ value }: SliderProps) {
           id="left"
           ref={arrowLeftRef}
           onClick={() => {
-            scrollBy(325, "left");
+            scrollLeftBy(325);
           }}
         >
           <img src={arrowLeft} alt="Arrow left" />
@@ -143,7 +62,7 @@ export function Slider({ value }: SliderProps) {
             />
 
             <div className="featured__product--description">
-              <h6 className="text-700 text-heading6">{product.title}</h6>
+              <h5 className="text-700 text-heading5">{product.title}</h5>
               <p className="text-caption">{product.caption}</p>
               <Link to="/reservation" className="text-caption link-underlined">
                 Make a reservation
@@ -157,7 +76,7 @@ export function Slider({ value }: SliderProps) {
           id="right"
           ref={arrowRightRef}
           onClick={() => {
-            scrollBy(325, "right");
+            scrollRightBy(325);
           }}
         >
           <img src={arrowRight} alt="Arrow right" />
