+++ src/scss/base/_typography.scss
@@ -18,10 +18,6 @@
   font-size: 1.5rem;
 }
 
-.text-heading6 {
-  font-size: 1.25rem;
-}
-
 .text-body {
   font-size: 1.25rem;
 }
@@ -37,8 +33,7 @@
 .text-heading2,
 .text-heading3,
 .text-heading4,
-.text-heading5,
-.text-heading6 {
+.text-heading5 {
   font-family: "stratos", sans-serif;
   font-weight: 600;
   line-height: 95%;
